code/requirements.txt
Flask
Flask-PyMongo
pymongo[tls,srv]
pytest
mockupdb
python-dotenv